clean the code to prepare for the implementation of the new functionnality
