+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,11 @@
         android:theme="@style/Theme.Bootcamp"
         tools:targetApi="31">
         <activity
+            android:name=".UserProfileCreationActivity"
+            android:exported="false" />
+        <activity
             android:name=".FirebaseBootcampActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -31,6 +34,7 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
