+++ app/src/main/java/com/github/drawyourpath/bootcamp/database/Database.kt
@@ -1,0 +1,7 @@
+package com.github.drawyourpath.bootcamp.database
+
+import android.widget.EditText
+
+abstract class Database {
+
+}
