+++ app/src/main/java/com/github/drawyourpath/bootcamp/database/FireDatabase.kt
@@ -1,0 +1,11 @@
+package com.github.drawyourpath.bootcamp.database
+
+import android.widget.EditText
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
+import java.util.concurrent.CompletableFuture
+
+class FireDatabase : Database() {
+
+}
