+++ app/src/main/java/com/github/drawyourpath/bootcamp/database/MockDataBase.kt
@@ -1,0 +1,7 @@
+package com.github.drawyourpath.bootcamp.database
+
+import android.widget.EditText
+
+class MockDataBase : Database() {
+    
+}
