+++ app/src/main/java/com/github/drawyourpath/bootcamp/userProfileCreation/UserProfileCreationActivity.kt
@@ -1,0 +1,12 @@
+package com.github.drawyourpath.bootcamp.userProfileCreation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.github.drawyourpath.bootcamp.R
+
+class UserProfileCreationActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_user_profile_creation)
+    }
+}
